+++ src/Domain/Twig/Service/TwigEngine.php
@@ -2,6 +2,9 @@
 
 namespace TotalCMS\Domain\Twig\Service;
 
+use Cake\Chronos\Chronos;
+use Cake\I18n\I18n;
+use Cake\I18n\RelativeTimeFormatter;
 use TotalCMS\Domain\Cache\Service\DevModeManager;
 use TotalCMS\Domain\License\Data\EditionFeature;
 use TotalCMS\Domain\License\Service\EditionFeatureService;
@@ -14,6 +17,7 @@ use Twig\Extension\CoreExtension;
 use Twig\Extension\DebugExtension;
 use Twig\Extension\StringLoaderExtension;
 use Twig\Extra\Html\HtmlExtension;
+use Twig\Extra\Intl\IntlExtension;
 use Twig\Extra\Markdown\MarkdownExtension;
 use Twig\Extra\Markdown\MarkdownRuntime;
 use Twig\Extra\String\StringExtension;
@@ -71,6 +75,20 @@ readonly class TwigEngine
 		$this->twig->addExtension(new StringLoaderExtension());
 		$this->twig->addExtension(new HtmlExtension());
 		$this->twig->addExtension(new MarkdownExtension());
+
+		// Configure locale for internationalization (requires intl extension)
+		if (extension_loaded('intl')) {
+			// Set PHP's default locale for IntlExtension and other intl functions
+			\Locale::setDefault($config->locale);
+			// Set CakePHP I18n locale for RelativeTimeFormatter translations
+			// Note: I18n::setLocale() internally calls \Locale::setDefault()
+			I18n::setLocale($config->locale);
+			$this->twig->addExtension(new IntlExtension());
+		}
+
+		// Configure Chronos to use locale-aware RelativeTimeFormatter
+		// This works without intl but translations will default to English
+		Chronos::diffFormatter(new RelativeTimeFormatter());
 
 		$this->twig->addRuntimeLoader(new class implements RuntimeLoaderInterface {
 			public function load(string $class)